Agro-copilot/backend/requirements.txt
fastapi==0.116.1
gunicorn==23.0.0
uvicorn==0.35.0
pydantic==2.11.7
numpy>=1.26
pillow>=10.0
torch>=2.2
torchvision>=0.17
transformers>=4.45
huggingface_hub>=0.24
safetensors>=0.4
python-dotenv>=1.0
